fix: move mfa functions to the base class
