+++ account-kit/signer/src/client/types.ts
@@ -208,7 +208,7 @@ export type SignerEndpoints = [
       stampedRequest: TSignedRequest;
       multiFactorType: MultiFactorType;
     };
-    Response: EnableMfaResult;
+    Response: AddMfaResult;
   },
   {
     Route: "/v1/auth-verify-multi-factor";
@@ -287,11 +287,11 @@ export type MfaFactor = {
 
 type MultiFactorType = "totp";
 
-export type EnableMfaParams = {
+export type AddMfaParams = {
   multiFactorType: MultiFactorType;
 };
 
-export type EnableMfaResult = {
+export type AddMfaResult = {
   multiFactorType: MultiFactorType;
   multiFactorId: string;
   multiFactorTotpUrl: string;
